requirements.txt
fastapi
uvicorn
openai
python-dotenv
azure-cosmos
